♻️ リファクタリング(interpreter.rs, lexical.rs): Float型をf64からf32に変更 ✨ 新機能(llvm_ir.rs): LLVM IR生成機能を追加 ♻️ リファクタリング(compiler.rs): main関数をeasy_compiler関数に分割し、コンパイルプロセスを整理

+++ core/src/lexical.rs
@@ -67,7 +67,7 @@ pub enum ValueType {
 #[derive(Debug, Clone, PartialEq)]
 pub enum Constant {
     Integer(i32),
-    Float(f64),
+    Float(f32),
 }
 
 /// トークン
@@ -131,7 +131,7 @@ impl Token
             _ if keyword.parse::<i64>().is_ok() =>
                 Some(Token::Constant(Constant::Integer(keyword.parse::<i32>().unwrap()))),
             _ if keyword.parse::<f64>().is_ok() =>
-                Some(Token::Constant(Constant::Float(keyword.parse::<f64>().unwrap()))),
+                Some(Token::Constant(Constant::Float(keyword.parse::<f32>().unwrap()))),
             _ => None,
         }
     }
